% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Retrieve data sets from the Statistics Estonia database.}
\usage{
get_data(id, language = "et")
}
\arguments{
\item{id}{Data set code.}

\item{language}{Language of the returned labels.
Either \code{et} or \code{en}.}
}
\description{
The function downloads the data set from the SDMX XML API.
}
\examples{
get_data("EH001")
get_data("EH001", language="en")
}
