% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_data.R
\name{search_data}
\alias{search_data}
\title{Search data sets from Statistics Estonia database.}
\usage{
search_data(phrase, language = "et")
}
\arguments{
\item{phrase}{Search phrase.}

\item{language}{Language of the search phrase and returned descriptions.
Either \code{et} or \code{en}.}
}
\description{
The function greps the phrase from the data sets' descriptions.
}
\examples{
search_data("produktiivsus")
search_data("alcohol", language="en")
}
